@@ -21,11 +213,25 @@
         "postinstall": "node ./node_modules/vscode/bin/install",
         "test": "node ./node_modules/vscode/bin/test"
     },
+    "dependencies": {
+        "shelljs": "^0.7.7",
+        "js-yaml": "^3.8.2",
+        "dockerfile-parse": "^0.2.0",
+        "k8s": "^0.4.12",
+        "tmp": "^0.0.31",
+        "pluralize": "^4.0.0",
+        "uuid": "^3.1.0"
+    },
     "devDependencies": {
-        "typescript": "^2.0.3",
+        "typescript": "^2.2.32",
+        "tslint": "^4.5.1",
         "vscode": "^1.0.0",
         "mocha": "^2.3.3",
         "@types/node": "^6.0.40",
         "@types/mocha": "^2.2.32"
+    },
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/Azure/vscode-kubernetes-tools"
     }
 }
